add maxAge for cookie, delete confirmation, withAuth utilities in routing

+++ public/js/deletePost.js
@@ -1,22 +1,26 @@
 const deletePostHandler = async (event) => {
     event.preventDefault();
     // console.log(event);
-
-    const id = window.location.toString().split('/')[
-        window.location.toString().split('/').length - 1
-      ];
-    
-    // console.log(id);
-    const response = await fetch (`/blog/edit/${id}`, {
-        method: "DELETE",
-    });
-    
-    if (response.ok) {
-        alert('Blogpost deleted.');
-        document.location.replace('/')
+    const confirmation = confirm('Really delete?')
+    if (confirmation) {
+        const id = window.location.toString().split('/')[
+            window.location.toString().split('/').length - 1
+        ];
+        
+        // console.log(id);
+        const response = await fetch (`/blog/edit/${id}`, {
+            method: "DELETE",
+        });
+        
+        if (response.ok) {
+            alert('Blogpost deleted.');
+            document.location.replace('/')
+        } else {
+            alert('Failed to delete blogpost');
+        };
     } else {
-        alert('Failed to delete blogpost');
-    };
+        return;
+    }
 };
 
 document
